test(omniconnect): add unit tests and fix lucide-react mock
- Implement comprehensive unit tests for `ConnectorRegistryImpl`.
- Add robust `lucide-react` mock in `__mocks__/lucide-react.ts` to prevent "missing export" errors in Vitest.
- Register the global mock in `tests/setup.ts`.
- remove `any` casts in tests to satisfy quality gates.
- Verified all 10 tests pass with `bun test`.

+++ src/lib/web3/entitlements.ts
@@ -302,7 +302,7 @@ async function checkNFTBalanceOnChain(
         args: [walletAddress],
       });
 
-      return balance;
+      return balance as bigint;
     } catch (error) {
       lastError = error as Error;
       console.error(`NFT balance check attempt ${attempt + 1} failed:`, error);
